feat: adjust fields for 'draw' and 'lottey' tables
The id of the lottey service, is now the one provided on the downloaded
datasets. Therefore all tables related with 'lottery' table, will be
related with the lottery service id.

Columns for the draw, can be null, since the table 'draw', can be used
for multiple lottery services (numbers can be different per service).